+++ cmd/commands/project.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"os"
 	"path/filepath"
 	"text/tabwriter"
 
@@ -45,7 +44,7 @@ func NewProjectCommand() *cobra.Command {
 		Short:   "Manage projects",
 		Run: func(cmd *cobra.Command, args []string) {
 			cmd.HelpFunc()(cmd, args)
-			os.Exit(1)
+			exit(1)
 		},
 	}
 
@@ -86,7 +85,7 @@ func NewProjectCreateCommand() *cobra.Command {
 
 # Create a new project in a specific path inside the GitOps repo
 
-  <BIN> project create <PROJECT_NAME> --installation-path path/to/bootstrap/root
+  <BIN> project create <PROJECT_NAME> --installation-path path/to/installation_root
 `),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if len(args) < 1 {